added piping

+++ server/server.js
@@ -22,7 +22,7 @@ var PORT = Number(process.env.PORT || 3000);
 var FILES_PATH = path.join(__dirname, '../files');
 var SPEED_TICK_TIME = 750; //ms
 //Init
-var oauth2ClientArray = {};
+var sessionData = {};
 var capture = false;
 var app = express();
 var server = http.createServer(app);
@@ -32,7 +32,9 @@ var torrents = {};
 var torrentObjs = {};
 var CLOUD = GDrive_1.GDrive;
 var filter = new Filter_1.Filter();
-var incognitoSessions = [];
+/**
+ *Calculates the percentage from given fraction, returns decimal
+ */
 function percentage(n) {
     var p = (Math.round(n * 1000) / 10);
     return (p > 100) ? 100 : p;
@@ -57,7 +59,7 @@ function saveToDriveHandler(sessionID, data) {
     }
     var stream = FILE.createReadStream(path.join(FILES_PATH, '../', obj.path));
     var cloudInstance = new CLOUD();
-    var req = cloudInstance.uploadFile(stream, obj.length, obj.mime, data.name, oauth2ClientArray[sessionID], false);
+    var req = cloudInstance.uploadFile(stream, obj.length, obj.mime, data.name, sessionData[sessionID].oauth2Client, false);
     cloudInstance.on('progress', function (data) {
         if (visitedPages[obj.id]) {
             visitedPages[obj.id].msg = "Uploaded " + percentage(data.uploaded / obj.length) + "%";
@@ -82,8 +84,8 @@ function saveToDriveHandler(sessionID, data) {
     });
 }
 /**
- *@params:  sessionID
-            data        {data:id}
+ *@param {sessionID}
+  @param {data} data data.id should be defined
  */
 function uploadDirToDrive(sessionID, data) {
     var id = data.id;
@@ -104,7 +106,7 @@ function uploadDirToDrive(sessionID, data) {
     }
     var dirSize = 0;
     var cloudInstance = new CLOUD();
-    cloudInstance.uploadDir(path.join(FILES_PATH, id), oauth2ClientArray[sessionID], false);
+    cloudInstance.uploadDir(path.join(FILES_PATH, id), sessionData[sessionID].oauth2Client, false);
     var uploaded = 0;
     cloudInstance.on("addSize", function (data) {
         dirSize = dirSize + data.size;
@@ -143,6 +145,9 @@ function clearVisitedPage(id) {
         if (visitedPages[id].progress == 100) {
             //  download completed but user requested to clear
             // delete downloaded file
+            if (visitedPages[id].isPiped) {
+                return;
+            }
             FILE.unlink(path.join(FILES_PATH, '../', visitedPages[id].path));
             delete visitedPages[id];
         }
@@ -153,6 +158,10 @@ function clearVisitedPage(id) {
         }
     }
 }
+/**
+ *Clears and deletes torrents
+ @param {id} id  The 'id' in 'torrents' object
+ */
 function clearTorrent(id) {
     if (!torrents[id].pinned) {
         io.emit("deleteKey", {
@@ -175,11 +184,11 @@ function clearTorrent(id) {
         }
     }
 }
-//TODO send pageVisited to its respective user using sessionID
 function middleware(data) {
     var sessionID = data.clientRequest.sessionID;
     var newFileName = null;
     if (filter.passed(data) && data.headers['content-length']) {
+        //check for duplicates
         var duplicates = Object.keys(visitedPages).filter(function (key) {
             return visitedPages[key].url == data.url;
         });
@@ -187,25 +196,50 @@ function middleware(data) {
             return false;
         }
         debug("DL:%s from %s", data.contentType, data.url);
+        //giving a name to our new file
         var uniqid = shortid.generate();
         var totalLength = data.headers['content-length'];
         var downloadedLength = 0;
         newFileName = uniqid + '.' + mime.extension(data.contentType);
         var completeFilePath = path.join(FILES_PATH, newFileName);
+        var piping = false;
         //create /files if it doesn't exist 
         if (!FILE.existsSync(FILES_PATH)) {
             FILE.mkdirSync(FILES_PATH);
         }
-        FILE.closeSync(FILE.openSync(completeFilePath, 'w')); //create an empty file
-        var stream = FILE.createWriteStream(completeFilePath);
-        data.stream.pipe(stream);
+        //check whether pipe to file or 
+        if (Object.keys(sessionData[sessionID].oauth2Client.credentials).length > 0 && sessionData[sessionID].piping) {
+            piping = true;
+            var cloudInstance = new CLOUD();
+            cloudInstance.uploadFile(data.stream, totalLength, data.contentType, newFileName, sessionData[sessionID].oauth2Client);
+            cloudInstance.on("fileUploaded", function (data) {
+                if (data.error) {
+                    visitedPages[uniqid].msg = data.error;
+                }
+                else {
+                    visitedPages[uniqid].msg = "Uploaded " + newFileName + " to GDrive.";
+                }
+                sendVisitedPagesUpdate(io, uniqid);
+            });
+            cloudInstance.on("progress", function (data) {
+                visitedPages[uniqid].msg = "Uploaded: " + percentage(data.uploaded / totalLength) + " %";
+                sendVisitedPagesUpdate(io, uniqid);
+            });
+        }
+        else {
+            FILE.closeSync(FILE.openSync(completeFilePath, 'w')); //create an empty file
+            var stream = FILE.createWriteStream(completeFilePath);
+            data.stream.pipe(stream);
+        }
         data.stream.on('data', function (chunk) {
             downloadedLength += chunk.length;
             var progress = percentage((downloadedLength / totalLength));
             if (visitedPages[uniqid]) {
                 if (visitedPages[uniqid].cleared) {
-                    stream.close();
-                    FILE.unlink(completeFilePath); //delete incomplete file
+                    if (stream) {
+                        stream.close();
+                        FILE.unlink(completeFilePath); //delete incomplete file
+                    }
                     delete visitedPages[uniqid];
                     io.emit('deleteKey', {
                         name: 'visitedPages',
@@ -258,6 +292,7 @@ function middleware(data) {
             pinned: false,
             progress: 0,
             length: data.headers['content-length'] * 1,
+            isPiped: piping,
             uploadTo: [] //holds list of session Ids to upload on dl complete
         };
         visitedPages[uniqid] = obj;
@@ -304,12 +339,13 @@ app.set("trust proxy", true);
 app.use(unblocker(middleware));
 app.use('/', express.static(path.join(__dirname, '../static')));
 app.use('/files', express.static(FILES_PATH));
-app.get('/', function (req, res) {
-    res.sendFile(path.join(__dirname, '../static', 'index.html'));
-});
 app.get('/oauthCallback', function (req, res) {
     var sessionID = req['sessionID'];
-    var oauth2Client = oauth2ClientArray[sessionID];
+    if (!sessionData[sessionID]) {
+        res.send('Invalid Attempt[E00]');
+        return false;
+    }
+    var oauth2Client = sessionData[sessionID].oauth2Client;
     if (!oauth2Client) {
         res.send('Invalid Attempt[E01]');
         return false;
@@ -333,21 +369,23 @@ app.get('/oauthCallback', function (req, res) {
 });
 // set up socket.io to use sessions
 io.use(function (socket, next) {
-    sessionMiddleware(socket.conn.request, socket.conn.request.res, next);
+    sessionMiddleware(socket.handshake, {}, next);
 });
 //handle socket.io connections
 io.on('connection', function (client) {
-    var sessionID = client.conn.request.sessionID;
-    if (!oauth2ClientArray[sessionID]) {
-        oauth2ClientArray[sessionID] = CLOUD.newOauthClient();
+    var sessionID = client.handshake.sessionID;
+    if (!sessionData[sessionID])
+        sessionData[sessionID] = { piping: false }; //init session
+    if (!sessionData[sessionID].oauth2Client) {
+        sessionData[sessionID].oauth2Client = CLOUD.newOauthClient();
     }
-    var consentPageUrl = CLOUD.getConsentPageURL(oauth2ClientArray[sessionID]);
+    var consentPageUrl = CLOUD.getConsentPageURL(sessionData[sessionID].oauth2Client);
     //Welcome new client
     client.emit('setObj', {
         name: 'status',
         value: {
             consentPageUrl: consentPageUrl,
-            logged: (Object.keys(oauth2ClientArray[sessionID].credentials).length > 0)
+            logged: (Object.keys(sessionData[sessionID].oauth2Client.credentials).length > 0)
         }
     });
     client.emit('setObj', {
@@ -357,6 +395,10 @@ io.on('connection', function (client) {
     client.emit('setObj', {
         name: 'torrents',
         value: torrents
+    });
+    client.emit('setObj', {
+        name: 'piping',
+        value: sessionData[sessionID].piping
     });
     client.on('clearVisitedPages', function () {
         Object.keys(visitedPages).forEach(function (id) {
@@ -415,7 +457,7 @@ io.on('connection', function (client) {
         var uniqid = shortid();
         torrentObjs[uniqid] = new Torrent_1.Torrent(data.magnet, FILES_PATH, uniqid);
         torrentObjs[uniqid].on("downloaded", function (path) {
-            //CLOUD.uploadDir(path, oauth2ClientArray[sessionID]);
+            //CLOUD.uploadDir(path, sessionData[sessionID].oauth2Client);
             torrents[uniqid].uploadTo.forEach(function (sessionId) {
                 uploadDirToDrive(sessionId, { id: uniqid });
             });
@@ -508,20 +550,15 @@ io.on('connection', function (client) {
             }
         });
     });
-    client.on("toggleIncognito", function () {
-        if (incognitoSessions.indexOf(sessionID) > -1) {
-            incognitoSessions.splice(incognitoSessions.indexOf(sessionID));
-        }
-        else {
-            incognitoSessions.push(sessionID);
-        }
+    client.on("togglePiping", function () {
+        sessionData[sessionID].piping = !sessionData[sessionID].piping;
     });
     client.on("uploadZipToCloud", function (data) {
         var id = data.id;
         var name = data.name;
         var loc = path.join(FILES_PATH, id + ".zip");
         var cloudInstance = new CLOUD();
-        cloudInstance.uploadFile(FILE.createReadStream(loc), torrents[id].length, mime.lookup(loc), name, oauth2ClientArray[sessionID], false);
+        cloudInstance.uploadFile(FILE.createReadStream(loc), torrents[id].length, mime.lookup(loc), name, sessionData[sessionID].oauth2Client, false);
         cloudInstance.on("progress", function (data) {
             torrents[id].msg = "Uploading Zip: " + percentage(data.uploaded / data.size) + "%";
             torrents[id].zipping = true;
